# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Homeroom API
  version: 0.1.0
  description: |
    Contract shared by the Next.js web app and the Flutter app. Every route is scoped to the
    caller's school and role by Postgres row-level security (see backend/migrations); the
    "Roles" line on each operation says who is meant to call it, and the database enforces it.

    **Conventions**
    - Auth: `Authorization: Bearer <jwt>`. The API maps the token to `people.id` and sets `app.person_id`.
    - IDs are UUIDs. Dates are `YYYY-MM-DD`; timestamps are RFC 3339 UTC. Money is a decimal string in GHS minor-unit-safe form (`"1850.00"`), never a float.
    - Lists are cursor-paginated: `?limit=50&cursor=...` returns `{ items, next_cursor }`.
    - Errors use `application/problem+json`. A row the caller may not see is reported as **404**, not 403, so existence is not leaked. 403 is for a visible resource the caller may not change.
    - Writes from the offline apps send an `Idempotency-Key` header; replays return the original result.
servers:
  - url: https://api.homeroom.example/v1
  - url: http://localhost:3000/api/v1
security:
  - bearer: []
...
paths:
...
  /auth/sessions:
    post:
      operationId: signInWithPassword
      tags: [ Auth ]
      summary: Sign in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ email, password ]
              properties: { email: { type: string, format: email }, password: { type: string } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { $ref: '#/components/responses/SignInResult' }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/otp:
    post:
      operationId: requestOtp
      tags: [ Auth ]
      summary: Request a one-time code by SMS (guardians without email)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ phone ]
              properties: { phone: { type: string, example: '+233240000001' } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '204': { description: 'Code sent (always 204, whether or not the number exists)' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [ Auth ]
      summary: Exchange a one-time code for a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ phone, code ]
              properties: { phone: { type: string }, code: { type: string, minLength: 6, maxLength: 6 } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { $ref: '#/components/responses/SignInResult' }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/sessions/current:
    delete:
      operationId: signOut
      tags: [ Auth ]
      summary: Sign out
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '204': { description: 'Signed out' }

  # ------------------------------------------------------------------ onboarding
...
  /auth/sessions/select:
    post:
      operationId: selectSessionPerson
      tags: [ Auth ]
      summary: Finish sign-in by choosing which role to continue as
      description: 'Used when sign-in returned selection_required because the account holds several people (for example an admin who is also a parent).'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ selection_token, person_id ]
              properties:
                selection_token: { type: string }
                person_id: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
...
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [ Auth ]
      summary: Exchange a refresh token for a new access and refresh token
      description: 'Refresh tokens rotate: each is single-use. Presenting one that was already used revokes the whole session family (token theft response).'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ refresh_token ]
              properties: { refresh_token: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
  /auth/password/forgot:
    post:
      operationId: requestPasswordReset
      tags: [ Auth ]
      summary: Email a password-reset link
      description: 'Always 204, whether or not the address has an account, so it cannot be used to discover accounts.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ email ]
              properties: { email: { type: string, format: email } }
      responses:
        '204': { description: 'Accepted' }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
  /auth/password/reset:
    post:
      operationId: resetPassword
      tags: [ Auth ]
      summary: Choose a new password from a reset or invitation link
      description: 'The token is single-use. Success signs the account out of every device.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ token, new_password ]
              properties:
                token: { type: string }
                new_password: { type: string, minLength: 8, maxLength: 200 }
      responses:
        '204': { description: 'Password changed' }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Session:
      description: Signed in
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Session' }
    SignInResult:
      description: Signed in, or a role choice is needed first
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SignInResult"
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Session:
      type: object
      required: [ access_token, refresh_token, expires_in, me ]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, description: 'Access token lifetime in seconds' }
        me: { $ref: '#/components/schemas/Me' }
    SelectionRequired:
      type: object
      required: [ selection_token, choices ]
      properties:
        selection_token: { type: string, description: 'Short-lived; exchange via /auth/sessions/select' }
        choices:
          type: array
          items:
            type: object
            required: [ person_id, role, school_name ]
            properties:
              person_id: { type: string, format: uuid }
              role: { $ref: '#/components/schemas/Role' }
              school_name: { type: string }
              full_name: { type: string }
    Contact:
      type: object
      properties:
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
      required: [ phone, email ]
...
    Problem:
      type: object
      required: [ status, title ]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }
        code: { type: string, example: term_closed }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
        request_id:
          type: string
          description: Matches the X-Request-Id response header and the server logs
...
    Role: { type: string, enum: [ admin, teacher, guardian, student ] }
    Me:
      type: object
      required: [ id, full_name, role, school, now ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        school:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            timezone: { type: string }
            term: { $ref: '#/components/schemas/Term' }
          required: [ id, name, timezone ]
        contact: { $ref: '#/components/schemas/Contact' }
        children: { type: array, items: { $ref: '#/components/schemas/StudentRef' }, description: 'Guardians only' }
        sections: { type: array, items: { $ref: '#/components/schemas/ClassSection' }, description: 'Teachers only' }
        now:
          type: string
          format: date-time
          description: The server's current time. Clients use it to correct a wrong device clock and to agree with the server about what 'today' is.
...
    StudentRef:
      type: object
      required: [ id, full_name ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
...
    Person:
      type: object
      required: [ id, full_name, role ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        contact: { type: object, properties: { phone: { type: string }, email: { type: string } } }
        assignments:
          type: array
          description: 'Teachers, in the admin directory: what they teach, in the current term'
          items:
            type: object
            properties:
              class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
              subject: { type: string }
            required: [ class_section, subject ]
...
    Term:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        closed: { type: boolean }
      required: [ id, name, starts_on, ends_on, closed ]
...
    ClassSection:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 6A }
        grade_level: { type: string }
        term_id: { type: string, format: uuid }
        homeroom_teacher: { $ref: '#/components/schemas/Person' }
        subject: { type: string, description: 'Present on /me and /class_sections for a teacher: the subject they teach here' }
      required: [ id, name, grade_level, term_id ]
...
    SignInResult:
      type: object
      description: Either a signed-in session, or (for an account holding several roles) the choices to pick from.
      properties:
        status:
          type: string
          enum:
            - signed_in
            - selection_required
        session:
          $ref: "#/components/schemas/Session"
        selection:
          $ref: "#/components/schemas/SelectionRequired"
      required: [ status ]